fix warnings
print.c:106:10: warning: variable ‘neg’ set but not used
tcpproxy.c:55:25: warning: cast to pointer from integer of different size
tcpproxy.c:64:7: warning: cast from pointer to integer of different size
testdelay.c:15:12: warning: cast from pointer to integer of different size
testdelay.c:16:32: warning: cast from pointer to integer of different size
testdelay.c:31:25: warning: cast to pointer from integer of different size

+++ tcpproxy.c
@@ -52,7 +52,7 @@ taskmain(int argc, char **argv)
 	fdnoblock(fd);
 	while((cfd = netaccept(fd, remote, &rport)) >= 0){
 		fprintf(stderr, "connection from %s:%d\n", remote, rport);
-		taskcreate(proxytask, (void*)cfd, STACK);
+		taskcreate(proxytask, (void*)(uintptr_t)cfd, STACK);
 	}
 }
 
@@ -61,7 +61,7 @@ proxytask(void *v)
 {
 	int fd, remotefd;
 
-	fd = (int)v;
+	fd = (int)(uintptr_t)v;
 	if((remotefd = netdial(TCP, server, port)) < 0){
 		close(fd);
 		return;
